+++ backend/src/app.js
@@ -17,6 +17,8 @@ app.use(express.json());
 app.use(cors(corsOption));
 app.use('/api', routes);
 
+require('./modules/publicAIP.js');
+
 const ErrorHandler = require('./middlewares/error.handler.middleware');
 app.use(ErrorHandler);
